requirements.txt
openai~=1.47.1
setuptools~=75.1.0
Flask
python-dotenv